main/LNN/requirements.txt
git+https://github.com/stanfordnmbl/osim-rl
numpy>=1.17.3
casadi==3.5.1
matplotlib>=3.1.1
tensorflow-gpu>=2.1.0
Torchdyn
Pytorch
Jax
Jaxlib

